ui/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pyyaml>=6.0
python-multipart>=0.0.9
